Fix macOS build

+++ descriptor/vehicle_desc.cc
@@ -375,7 +375,7 @@ void vehicle_desc_t::calc_checksum(checksum_t *chk) const
 	chk->input(len);
 	chk->input(leader_count);
 	chk->input(trailer_count);
-	chk->input(engine_type);
+	chk->input((uint8)engine_type);
 	// freight
 	const xref_desc_t *xref = get_child<xref_desc_t>(2);
 	chk->input(xref ? xref->get_name() : "NULL");
@@ -395,7 +395,7 @@ void vehicle_desc_t::calc_checksum(checksum_t *chk) const
 	chk->input(upgrades);
 	chk->input(is_tilting ? 1 : 0);
 	chk->input(mixed_load_prohibition ? 1 : 0);
-	chk->input(override_way_speed ? 1 : 0); 
+	chk->input(override_way_speed ? 1 : 0);
 	chk->input(basic_constraint_prev);
 	chk->input(basic_constraint_next);
 	chk->input(way_constraints.get_permissive());
